feat: id를 int 타입으로 변경

+++ Server/src/main/java/cooperativedistributedsystem/whiteboard/EnterController.java
@@ -11,8 +11,8 @@ import java.util.Set;
 public class EnterController {
 
     private final MessageRepository messageRepository;
-    private long objectId = 0L;
-    private final long interval = 1024L;
+    private int objectId = 0;
+    private final int interval = 1024;
 
     @GetMapping
     public Set<Message> enter() {
@@ -20,12 +20,12 @@ public class EnterController {
     }
 
     @GetMapping("id")
-    public long allocateObjectId() {
+    public int allocateObjectId() {
         return getObjectStartId();
     }
 
-    private synchronized long getObjectStartId() {
-        long startId = objectId;
+    private synchronized int getObjectStartId() {
+        int startId = objectId;
         objectId += interval;
         return startId;
     }
